requirements.txt
# API 相關
requests==2.31.0
aiohttp==3.9.3

# 資料庫
pymongo>=4.6.0
motor>=3.3.0
mongomock>=4.1.2

# 時間處理
pytz==2024.1

# 非同步支援
asyncio==3.4.3

# 日誌和監控
python-json-logger>=2.0.7

# 測試相關
pytest==8.1.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock>=3.12.0
pytest-timeout>=2.2.0
coverage>=7.4.0
freezegun>=1.4.0  # 用於時間凍結測試

# 程式碼品質
black>=24.1.0
flake8>=7.0.0
mypy==1.9.0
isort>=5.13.0

# 文件生成
sphinx>=7.2.0
sphinx-rtd-theme>=2.0.0
autodoc>=0.5.0

# 開發工具
pre-commit>=3.6.0
tox>=4.13.0
python-dotenv==1.0.1

# 新增的套件
schedule>=1.2.1

# 安全性工具
bandit>=1.7.0
safety>=3.0.0

# 型別檢查
types-requests==2.31.0.20240311
types-pytz==2024.1.0.20240203

# 日誌處理
colorlog==6.8.2

# 新增的套件
openai==1.68.2
PyYAML==6.0.1
python-dateutil==2.9.0 